Renamed files for site, template and pnp modules
