+++ playbooks/device_details.yml
@@ -1,0 +1,67 @@
+template_details:
+  - proj_name: "Onboarding Configuration"
+    device_config: "hostname cat9k-1\n"
+    language: "velocity"
+    family: "Switches and Hubs"
+    type: "IOS-XE"
+    variant: "XE"
+    temp_name: "temp_cat9k-1"
+    description: "Test Template 1"
+  - proj_name: "Onboarding Configuration"
+    device_config: "hostname cat9k-2\n"
+    language: "velocity"
+    family: "Switches and Hubs"
+    type: "IOS-XE"
+    variant: "XE"
+    temp_name: "temp_cat9k-2"
+    description: "Test Template 2"
+  - proj_name: "Onboarding Configuration"
+    device_config: "hostname cat9k-3\n"
+    language: "velocity"
+    family: "Switches and Hubs"
+    type: "IOS-XE"
+    variant: "XE"
+    temp_name: "temp_cat9k-3"
+    description: "Test Template 3"
+
+device_details:
+  - site_type: "building"
+    bld_lat: 12.93570689258
+    bld_long: 80.24992245822832
+    bld_name: "Trill"
+    site_name: "Global/Chennai"
+    image_name: "cat9k_iosxe.17.04.01.SPA.bin"
+    proj_name: "Onboarding Configuration"
+    temp_name: "temp_cat9k-1"
+    device_version: "2"
+    device_number: "AB2425L8M7"
+    device_name: "Cat9k-1"
+    device_state: "Unclaimed"
+    device_id: "C9300-25UX"
+  - site_type: "building"
+    bld_lat: 12.93570689258
+    bld_long: 80.24992245822832
+    bld_name: "Trill"
+    site_name: "Global/Chennai"
+    image_name: "cat9k_iosxe.17.04.01.SPA.bin"
+    proj_name: "Onboarding Configuration"
+    temp_name: "temp_cat9k-2"
+    device_version: "2"
+    device_number: "CD2425L8M7"
+    device_name: "Cat9k-2"
+    device_state: "Unclaimed"
+    device_id: "C9300-25UX"
+  - site_type: "building"
+    bld_lat: 12.93570689258
+    bld_long: 80.24992245822832
+    bld_name: "Trill"
+    site_name: "Global/Chennai"
+    image_name: "cat9k_iosxe.17.04.01.SPA.bin"
+    proj_name: "Onboarding Configuration"
+    temp_name: "temp_cat9k-3"
+    device_version: "2"
+    device_number: "EF2425L8M7"
+    device_name: "Cat9k-3"
+    device_state: "Unclaimed"
+    device_id: "C9300-25UX"
+
