+++ tests/unit/modules/dnac/fixtures/pnp_intent.json
